Refactoring with new api path payment-methods/:id/fee/calculate
