+++ src/main/java/it/pagopa/ecommerce/payment/methods/application/PaymentMethodService.java
@@ -1,5 +1,6 @@
 package it.pagopa.ecommerce.payment.methods.application;
 
+import it.pagopa.ecommerce.payment.methods.client.AfmClient;
 import it.pagopa.ecommerce.payment.methods.domain.aggregates.PaymentMethod;
 import it.pagopa.ecommerce.payment.methods.domain.aggregates.PaymentMethodFactory;
 import it.pagopa.ecommerce.payment.methods.domain.valueobjects.PaymentMethodAsset;
@@ -12,8 +13,13 @@ import it.pagopa.ecommerce.payment.methods.domain.valueobjects.PaymentMethodType
 import it.pagopa.ecommerce.payment.methods.exception.PaymentMethodNotFoundException;
 import it.pagopa.ecommerce.payment.methods.infrastructure.PaymentMethodDocument;
 import it.pagopa.ecommerce.payment.methods.infrastructure.PaymentMethodRepository;
+import it.pagopa.ecommerce.payment.methods.server.model.BundleOptionDto;
+import it.pagopa.ecommerce.payment.methods.server.model.PaymentMethodStatusDto;
+import it.pagopa.ecommerce.payment.methods.server.model.PaymentOptionDto;
+import it.pagopa.ecommerce.payment.methods.server.model.TransferDto;
 import it.pagopa.ecommerce.payment.methods.utils.ApplicationService;
 import it.pagopa.ecommerce.payment.methods.utils.PaymentMethodStatusEnum;
+import it.pagopa.generated.ecommerce.gec.v1.dto.TransferListItemDto;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.util.Pair;
@@ -21,20 +27,32 @@ import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 import java.util.stream.Collectors;
+
+import static java.util.stream.Collectors.toList;
 
 @Service
 @ApplicationService
 @Slf4j
 public class PaymentMethodService {
 
-    @Autowired
+    private final AfmClient afmClient;
+
     private PaymentMethodRepository paymentMethodRepository;
 
-    @Autowired
     private PaymentMethodFactory paymentMethodFactory;
+
+    @Autowired
+    public PaymentMethodService(
+            AfmClient afmClient,
+            PaymentMethodRepository paymentMethodRepository,
+            PaymentMethodFactory paymentMethodFactory
+    ) {
+        this.afmClient = afmClient;
+        this.paymentMethodFactory = paymentMethodFactory;
+        this.paymentMethodRepository = paymentMethodRepository;
+    }
 
     public Mono<PaymentMethod> createPaymentMethod(
                                                    String paymentMethodName,
@@ -146,6 +164,95 @@ public class PaymentMethodService {
                 .map(this::docToAggregate);
     }
 
+    public Mono<it.pagopa.ecommerce.payment.methods.server.model.BundleOptionDto> computeFee(
+                                                                                             Mono<PaymentOptionDto> paymentOptionDto,
+                                                                                             String paymentMethodId,
+                                                                                             Integer maxOccurrences
+    ) {
+        return paymentMethodRepository.findById(paymentMethodId)
+                .switchIfEmpty(Mono.error(new PaymentMethodNotFoundException(paymentMethodId)))
+                .flatMap(
+                        pm -> paymentOptionDto.map(
+                                po -> new it.pagopa.generated.ecommerce.gec.v1.dto.PaymentOptionDto()
+                                        .bin(po.getBin())
+                                        .paymentAmount(po.getPaymentAmount())
+                                        .idPspList(po.getIdPspList())
+                                        .paymentMethod(pm.getPaymentMethodTypeCode())
+                                        .primaryCreditorInstitution(po.getPrimaryCreditorInstitution())
+                                        .touchpoint(po.getTouchpoint())
+                                        .transferList(
+                                                po.getTransferList()
+                                                        .stream()
+                                                        .map(
+                                                                t -> new TransferListItemDto()
+                                                                        .creditorInstitution(t.getCreditorInstitution())
+                                                                        .digitalStamp(t.getDigitalStamp())
+                                                                        .transferCategory(t.getTransferCategory())
+                                                        )
+                                                        .collect(toList())
+                                        )
+
+                        ).flatMap(reqBody -> afmClient.getFees(reqBody, maxOccurrences))
+                                .map(bo -> {
+                                    bo.setBundleOptions(
+                                            removeDuplicatePsp(bo.getBundleOptions())
+                                    );
+                                    return bo;
+                                })
+                                .map(bo -> bundleOptionToResponse(bo, pm.getPaymentMethodStatus()))
+
+                );
+
+    }
+
+    private List<it.pagopa.generated.ecommerce.gec.v1.dto.TransferDto> removeDuplicatePsp(
+                                                                                          List<it.pagopa.generated.ecommerce.gec.v1.dto.TransferDto> transfers
+    ) {
+        return transfers
+                .stream()
+                .collect(
+                        Collectors.collectingAndThen(
+                                Collectors.toCollection(
+                                        () -> new TreeSet<>(
+                                                Comparator.comparing(
+                                                        it.pagopa.generated.ecommerce.gec.v1.dto.TransferDto::getIdPsp
+                                                )
+                                        )
+                                ),
+                                ArrayList::new
+                        )
+                );
+    }
+
+    private BundleOptionDto bundleOptionToResponse(
+                                                   it.pagopa.generated.ecommerce.gec.v1.dto.BundleOptionDto bundle,
+                                                   String paymentMethodStatus
+    ) {
+        return new it.pagopa.ecommerce.payment.methods.server.model.BundleOptionDto()
+                .belowThreshold(bundle.getBelowThreshold())
+                .paymentMethodStatus(PaymentMethodStatusDto.valueOf(paymentMethodStatus))
+                .bundleOptions(
+                        bundle.getBundleOptions() != null ? bundle.getBundleOptions()
+                                .stream()
+                                .map(
+                                        t -> new TransferDto()
+                                                .abi(t.getAbi())
+                                                .bundleDescription(t.getBundleDescription())
+                                                .bundleName(t.getBundleName())
+                                                .idBrokerPsp(t.getIdBrokerPsp())
+                                                .idBundle(t.getIdBundle())
+                                                .idChannel(t.getIdChannel())
+                                                .idCiBundle(t.getIdCiBundle())
+                                                .idPsp(t.getIdPsp())
+                                                .onUs(t.getOnUs())
+                                                .paymentMethod(t.getPaymentMethod())
+                                                .primaryCiIncurredFee(t.getPrimaryCiIncurredFee())
+                                                .taxPayerFee(t.getTaxPayerFee())
+                                                .touchpoint(t.getTouchpoint())
+                                ).toList() : new ArrayList<>()
+                );
+    }
+
     private PaymentMethod docToAggregate(PaymentMethodDocument doc) {
         if (doc == null) {
             return null;
@@ -163,4 +270,5 @@ public class PaymentMethodService {
                 new PaymentMethodAsset(doc.getPaymentMethodAsset())
         );
     }
+
 }
